4.15-4.21: tokenValidation middleware and delete blogs only by theirauthors

+++ part4/blog_list/controllers/blog.js
@@ -1,27 +1,58 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const notes = await Blog
+  const blogs = await Blog
     .find({}).populate('user', { username: 1, name: 1 })
 
-  response.json(notes)
+  response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const user = await User.findById(request.body.userId)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) { 
+    return response.status(401).json({
+      error: 'invalid token'
+    })
+  }
+  const user = await User.findById(decodedToken.id)
+
   const blog = new Blog({...request.body, user: user._id, })
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
-  response.json(savedBlog)
+  response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
-
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({
+      error: 'invalid token'
+    })
+  }
+  const blog = await Blog.findById(request.params.id)
+  console.log("BLOG:", blog)
+  if (blog === null) {
+    response.status(400).json({
+      error: 'Could not find a blog with this id.'
+    })
+    return
+  }
+  const user = await User.findById(decodedToken.id)
+  console.log("USER:", user.id)
+  console.log("BLOG USER:", blog.user)
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+    return
+  }
+  return response.status(400).json({
+    error: 'You cannot delete another user\'s blog.'
+  })
+  // response.status(400).end()
 })
 
 blogsRouter.put('/:id', async (request, response) => {
